requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
groq==0.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
httpx==0.26.0
pydantic==2.5.3
PyPDF2==3.0.1
numpy==1.26.4
